org cem bitmap
